requirements.txt
pdfplumber
python-docx
scikit-learn
rapidfuzz
streamlit
